Support wider range of behat/gherkin versions
The TableNode was refactored in 3.0 to become an Immutable
data object where the full table must be passed as an array in
the constructor.

Update the PaddedTableNode to support constructing either
a v2.x or v3+ TableNode so that we can support a wider range
of gherkin versions to fit in with what is already in the end-user
project.