Don't fail when HTML contains valid unclosed tags eg <input>
The simple_xml parser expects actual XML and fails if the table
contains unclosed HTML tags like <input> etc. In earlier versions
of the mink browserkit driver, the HTML provided was in fact an
XML export of the node (so closing / self-closing tags were added
by mink) but since 1.3.0 (minkphp/MinkBrowserKitDriver#93) we
correctly get the actual HTML rendered by the server.

This does unfortunately mean the parser is much more tolerant of
bad HTML now - but the only solution for that would be to add a
full HTML purifier or similar library which feels too heavy for
a required dependency.